requirements.txt
langgraph 
langchain_openai 
langchain_community 
langchain_core 
tavily-python 
wikipedia
langchain-anthropic
langchain-google-genai
streamlit
chromadb
llama-index-embeddings-huggingface